update name and timestamp
